Utils does not depend anymore on biggraphite modules.
Three classes have been extracted:
- settings
- accessor_factory
- cache_factory
