+++ biggraphite/utils.py
@@ -14,69 +14,17 @@
 # limitations under the License.
 """Utility module."""
 
-import os
 import logging
-import distutils
+import os
+
 import prometheus_client
 
-from biggraphite.drivers import cassandra as bg_cassandra
-from biggraphite.drivers import elasticsearch as bg_elasticsearch
-from biggraphite.drivers import hybrid as bg_hybrid
-from biggraphite.drivers import memory as bg_memory
-from biggraphite import metadata_cache
 
-log = logging.getLogger(__name__)
-
-DRIVERS = frozenset([
-    ("cassandra", bg_cassandra),
-    ("elasticsearch", bg_elasticsearch),
-    ("memory", bg_memory),
-])
-CACHES = frozenset([
-    ("disk", metadata_cache.DiskCache),
-    ("memory", metadata_cache.MemoryCache),
-])
-
-
-def strtoint(value):
-    """Cast a string to an integer."""
-    if value is None:
-        return None
-    return int(value)
-
-
-def strtobool(value):
-    """Cast a string to a bool."""
-    if value is None:
-        return None
-    if type(value) is bool:
-        return value
-    return distutils.util.strtobool(value)
-
-
-def strvalidator(value):
-    """Cast a value to string."""
-    if value is None:
-        return None
-    return str(value)
-
-
-DEFAULT_DRIVER = "cassandra"
-DEFAULT_CACHE = "memory"
 DEFAULT_LOG_LEVEL = "WARNING"
 DEFAULT_ADMIN_PORT = None
-OPTIONS = {
-    "driver": str,
-    "metadata_driver": strvalidator,
-    "data_driver": strvalidator,
-    "cache": str,
-    "cache_size": strtoint,
-    "cache_ttl": strtoint,
-    "cache_sync": strtobool,
-    "loglevel": str,
-    "storage_dir": str,
-    "admin_port": strtoint,
-}
+
+
+log = logging.getLogger(__name__)
 
 
 class Error(Exception):
@@ -106,120 +54,6 @@ def start_admin(settings):
 start_admin.started = False
 
 
-def accessor_from_settings(settings):
-    """Get Accessor from configuration.
-
-    Args:
-      settings: dict(str -> value).
-
-    Returns:
-      Accessor (not connected).
-    """
-    driver_name = settings.get('driver', DEFAULT_DRIVER)
-    metadata_driver = settings.get('metadata_driver', None)
-    data_driver = settings.get('data_driver', None)
-
-    if metadata_driver is None and data_driver is None:
-        return _build_simple_accessor(driver_name, settings)
-    else:
-        if metadata_driver is None:
-            raise ConfigError("Metadata driver is not provided. Please specify --metadata_driver")
-        if data_driver is None:
-            raise ConfigError("Data driver is not provided. Please specify --data_driver")
-
-        metadata_accessor = _build_simple_accessor(metadata_driver, settings)
-        data_accessor = _build_simple_accessor(data_driver, settings)
-
-        return bg_hybrid.HybridAccessor(
-            "%s_%s" % (metadata_driver, data_driver),
-            metadata_accessor,
-            data_accessor
-        )
-
-
-def _build_simple_accessor(driver_name, settings):
-    driver_settings = {}
-
-    # Get driver specific settings.
-    prefix = driver_name + '_'
-    for key, value in settings.items():
-        if key.startswith(prefix):
-            key = key[len(prefix):]
-            driver_settings[key] = value
-
-    for name, driver in DRIVERS:
-        if name == driver_name:
-            return driver.build(**driver_settings)
-
-    raise ConfigError("Invalid driver '%s'." % driver_name)
-
-
-def cache_from_settings(accessor, settings, cname=None):
-    """Get Cache from configuration.
-
-    Args:
-      settings: dict(str -> value).
-
-    Returns:
-      Cache (not opened).
-    """
-    cache_name = settings.get('cache', DEFAULT_CACHE)
-    cache_settings = {
-        'path': settings.get('storage_dir'),
-    }
-    for opt in ['size', 'ttl', 'sync']:
-        value = settings.get('cache_%s' % opt)
-        if value is not None:
-            cache_settings[opt] = value
-
-    for name, cache in CACHES:
-        if name == cache_name:
-            return cache(accessor, cache_settings, cname)
-
-    raise ConfigError("Invalid value '%s' for BG_CACHE." % cache_name)
-
-
-def add_argparse_arguments(parser):
-    """Add generic BigGraphite arguments to an argparse parser.
-
-    Args:
-      parser: argparse.ArgumentParser()
-    """
-    parser.add_argument(
-        "--driver",
-        help="BigGraphite driver (%s)" % ', '.join([v[0] for v in DRIVERS]),
-        default=DEFAULT_DRIVER)
-    parser.add_argument(
-        "--metadata_driver",
-        help="BigGraphite metadata driver (%s)" % ', '.join([v[0] for v in DRIVERS]))
-    parser.add_argument(
-        "--data_driver",
-        help="BigGraphite data driver (%s)" % ', '.join([v[0] for v in DRIVERS]))
-    parser.add_argument(
-        "--cache",
-        help="BigGraphite cache (%s))" % ', '.join([v[0] for v in CACHES]),
-        default=DEFAULT_CACHE)
-    parser.add_argument(
-        "--cache-size",
-        help="Metadata cache size.")
-    parser.add_argument(
-        "--cache-sync",
-        help="Metadata cache sync.")
-    parser.add_argument(
-        "--storage_dir", metavar="PATH",
-        help="Storage path (cache, etc..)")
-    parser.add_argument(
-        "--loglevel", metavar="LEVEL",
-        help="Log level (DEBUG, INFO, WARNING, ERROR, CRITICAL)",
-        default=DEFAULT_LOG_LEVEL)
-    parser.add_argument(
-        "--admin_port", metavar="PORT",
-        help="Admin port with /metrics",
-        default=DEFAULT_ADMIN_PORT)
-    bg_cassandra.add_argparse_arguments(parser)
-    bg_elasticsearch.add_argparse_arguments(parser)
-
-
 def set_log_level(settings):
     """Set logs level according to settings."""
     logger = logging.getLogger('biggraphite')
@@ -230,62 +64,6 @@ def set_log_level(settings):
             "%(asctime)s [%(levelname)s] %(name)s: %(message)s"))
         logger.addHandler(handler)
     logger.setLevel(settings.get("loglevel", DEFAULT_LOG_LEVEL))
-
-
-def get_setting(settings, name):
-    """Get a specific setting from Carbon/Django like settings."""
-    res = None
-    found = False
-    try:
-        res = settings[name]
-        found = True
-    except (TypeError, KeyError):
-        try:
-            res = getattr(settings, name)
-            found = True
-        except Exception:
-            pass
-    return res, found
-
-
-def settings_from_args(args):
-    """Create settings dict from args.
-
-    Args:
-      args: argparse.Namespace, parsed arguments.
-
-    Returns:
-      dict(string: value): settings
-    """
-    return settings_from_confattr(args, prefix="")
-
-
-def settings_from_confattr(conf, prefix="bg_"):
-    """Create settings dict from Django/Carbon like settings.
-
-    Args:
-      settings: either carbon_conf.Settings or a Django-like settings object
-
-    Returns:
-      dict(string: value): settings
-    """
-    settings = {}
-
-    options = dict(OPTIONS)
-    for name, driver in DRIVERS:
-        options.update(
-            {('%s_' % name) + k: v for k, v in driver.OPTIONS.items()})
-
-    for option, validator in options.items():
-        option_u = (prefix + option).upper()
-        option_l = (prefix + option).lower()
-        value, found = get_setting(conf, option_u)
-        if not found:
-            value, found = get_setting(conf, option_l)
-        if found:
-            settings[option] = validator(value)
-
-    return settings
 
 
 def manipulate_paths_like_upstream(_executable, sys_path):
@@ -309,3 +87,13 @@ def setup_graphite_root_path(carbon_util_file):
     if os.path.dirname(carbon_util_file) == "/opt/graphite/lib/carbon":
         if "GRAPHITE_ROOT" not in os.environ:
             os.environ["GRAPHITE_ROOT"] = "/opt/graphite"
+
+
+def round_down(rounded, divider):
+    """Round down an integer to a multiple of divider."""
+    return int(rounded) // divider * divider
+
+
+def round_up(rounded, divider):
+    """Round up an integer to a multiple of divider."""
+    return int(rounded + divider - 1) // divider * divider
