+++ tests/cli/test_command_read.py
@@ -18,7 +18,7 @@ import unittest
 import argparse
 
 from biggraphite.cli import command_read
-from biggraphite import utils as bg_utils
+from biggraphite import settings as bg_settings
 from biggraphite import accessor as bg_accessor
 from tests import test_utils as bg_test_utils
 
@@ -34,7 +34,7 @@ class TestCommandRead(bg_test_utils.TestCaseWithFakeAccessor):
         cmd = command_read.CommandRead()
 
         parser = argparse.ArgumentParser()
-        bg_utils.add_argparse_arguments(parser)
+        bg_settings.add_argparse_arguments(parser)
         cmd.add_arguments(parser)
         opts = parser.parse_args([name])
         cmd.run(self.accessor, opts)
