+++ tests/cli/test_command_syncdb.py
@@ -18,7 +18,7 @@ import unittest
 import argparse
 
 from biggraphite.cli import command_syncdb
-from biggraphite import utils as bg_utils
+from biggraphite import settings as bg_settings
 from tests import test_utils as bg_test_utils
 
 
@@ -28,7 +28,7 @@ class TestCommandSyncdb(bg_test_utils.TestCaseWithFakeAccessor):
         cmd = command_syncdb.CommandSyncdb()
 
         parser = argparse.ArgumentParser()
-        bg_utils.add_argparse_arguments(parser)
+        bg_settings.add_argparse_arguments(parser)
         cmd.add_arguments(parser)
         opts = parser.parse_args(['--retention', '86400*1s:10080*60s'])
         cmd.run(self.accessor, opts)
