+++ tests/cli/test_command_web.py
@@ -23,7 +23,7 @@ except ImportError:
     gourde = None
 
 from biggraphite.cli import command_web
-from biggraphite import utils as bg_utils
+from biggraphite import settings as bg_settings
 from tests import test_utils as bg_test_utils
 
 
@@ -41,7 +41,7 @@ class TestCommandWeb(bg_test_utils.TestCaseWithFakeAccessor):
         cmd = command_web.CommandWeb()
 
         parser = argparse.ArgumentParser()
-        bg_utils.add_argparse_arguments(parser)
+        bg_settings.add_argparse_arguments(parser)
         cmd.add_arguments(parser)
         opts = parser.parse_args(["--dry-run"])
         cmd.run(self.accessor, opts)
